Update LICENSE
Here's the deal. When @stanhebben told me the license of the repo, he said that ZenScript was MIT and I got confused and thought the whole repo was MIT, it it not. After speaking to people, due to the original MIT licensing being invalid, changing the license now to a valid license is legal.

This license has been approved by Stan