Adding recipes.replaceAllOccurences(IIngredient,IIngredient) as a new zenscript command.
It will search through all vanilla style recipes for any recipe containing the first ingredient, remove that recipe, and create a new recipe that is identical, except the second ingredient has been substituted for the first.

I feel this will be useful in cases such as replacing all Machine Frames from one mod to another, or (the case that made me write this) switching all Industrial Foregoing recipes to use Pneumaticraft Plastic instead of IF Plastic.