Options for better enchanted book creation: makeBook, makeBookTag, etc.
This extrapolates the process of making the DataMap for each individual
enchant out into a Map<String, IData> function which can then be sent to
the DataMap. This reduces the amount of code duplication. makeTag and
makeBookTag now use this function to create maps with the key "ench" and
"StoredEnchantments" respectively.

makeBook specifically uses makeNBTInternal to create an instance of
<minecraft:enchanted_book> with the relevant NBT applied to it. This
means that functionally the two examples below are identical:

```
var ench = <enchantment:minecraft:protection>.makeEnchantment(3);

var ench_book = ench.makeBook();
var ench_book2 = <minecraft:enchanted_book>.withTag(ench_book.makeBookTag());
```

Previously using the default `makeTag` would result in a book that was
**enchanted** rather than a book that had stored enchantments on it.

---

Additionally adds three convenient functions to itemUtils:
createEnchantedBook, which takes a list of IEnchantment and then creates
a book; combineEnchantments, which takes (optionally) a base key
(default is "ench"), and a list of IEnchantment.

The latter function doesn't check to make sure that the string options
are valid (i.e., one of "ench", "StoredEnchantments") so there is a
potential failure point there.