POC: ItemTransformers (see desc for more)
Transformers implemente the interface IIngredientTransformer<T extends IIngredient>
Honestly I don't even remember why I made it generic :thinking:

They need to have a serializer, which also needs to be registered to CrafttweakerRegistries.REGISTRY_TRANSFORMER_SERIALIZER

For custom Transformers, there is IIngredient.transformCustom(uid, function)
The uid is a uid for the transformer and the only thing that will be sent across the wire.
The function can be null, but only of another call with the same uid sets it.
In short, at least one call per uid must not be empty (if multiple aren't then there may be desyncs between client and server, since the client will receive the last registered function, whereas the server will use whatever it got (TODO: change this?)