Merged dumpBracket and dump x Sources, Added CommandCollectionEvent
The command Collection Event is called after the CT commands were initialized but before they are added to the dispatcher.
This allows you to add subcommands properly.
Only dump commands added by the `@BracketDumper` (see below) won't be
present already because we add them later to spot conflicting
dump/subcommand names.

Now the `/ct dump xxx` command will also use `@BracketDumper` methods as sources
That was we don't need two handlers that do the same, one for
dumpBrackets and one for the dump command.
For this, the internal structure of how BracketDumpers were created was changed.
The annotation was changed to allow for changes in the generated
filename and Display Name (for the help command)

The rest of the commit is applying the changes to the existing command
dumpers