[1.16] Implement Global Loot Modifiers (#1092)
* Add global loot modifier support

Support for built-in Loot Conditions is still not done, along with some
elements of the loot context.

There is also some duplication with the wrappers and the classes, and
documentation. Nevertheless this is still a good-enough start

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Account for scripts loading on the client

Global Loot Modifiers don't get loaded on the client, and the map is
never initialized. This would cause a crash, but now the exception is
caught and logged.

We could bypass the need via an injected getter and an explicit null
check, but the amount of times this will happen is minuscule compared to
the amount of checks. Basically, the amount of queries is usually way
less than the amount of registrations and removals.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Use null object instead of raw nulls in wrapper

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Avoid full blown server crash in case of mistakes in GLM body

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Split modifier manager from loot manager

This proves useful in case other mods may want to ZenExpand `loot` with
their own additions or if CraftTweaker wants to add some additional
functionalities, like real-time querying of loot tables.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Expand loot context with damage source, origin, id

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Provide bracket handler to obtain Enchantment

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Add builders for vanilla-provided loot conditions

Some builders are still missing, those are marked with TODOs on the
code. Some particular care has to be taken in this case due to how
convoluted vanilla makes those, with lots and lots of moving pieces.

These builders represent a way to make building conditions easier for
players, since functions are provided for them already, using a simple
builder pattern. The core of this functionality is
CTLootConditionBuilder, which provides an entry point for all the other
builders, allowing for chaining.

A JSON Loot Condition is also provided so that people can leverage IData
if they want more control (for whatever control JSON provides) or want
to integrate with other mod's loot conditions.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Add helper method for builder-less loot conditions

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Add helper method for builder-less loot conditions

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Add full support for vanilla loot tables... sorry Jared

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Move to Access Transformers instead of Method Handles

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Split off predicates into their own package

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Refactor Loot Modifiers in their own package

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Add some common loot modifiers

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Add **HUGE BODGE** for Nullable types

I repeat **HUGE BODGE**

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Add example script

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* YASSSS

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Add some helper functions

TODO: Figure out how not to cause documentation to be overwritten

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Move to automatic registration of condition builders

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Make min/max methods preserve previous bounds and general cleanup

Also add damage-related helpers for items

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Add dump command for all registered modifiers

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Add example for extension methods

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Additional code fixes

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Fix Alternative logic not matching vanilla

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Provide additional conditions for basic logic

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Give meaning to createInAnd and createInOr + fixes

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Move condition builder registry into own class

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Change documentation location

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Complete set of logic gates with NOT condition

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* Move naming checks to NameUtils

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>

* The Push:tm: - Documentation for everything

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>