Replace Mixin with MethodHandle collection
The Method Handle solution acts exactly the same as the Mixin solution
while also removing the need for bytecode injection into the target.

The usage of reflection to identify the target field via its type
ensures compatibility across compilers, which may choose a different
name for the synthetic field over `val$bimap`. The type of the field, on
the other hand, is guaranteed to be `BiMap`.

Another upside of the usage of method handles is that this ensures that
the map is replaced and checked only if needed, avoiding useless
`instanceof` checks and new object creation if not needed. On the
contrary, checks are now made only when a tag needs to be added and are
specific to the manager. A point of concern may be the presence of the
check even after the first time, when the map has been transformed. An
alternative to this may be exception bubbling, so assuming the map is
correctly made mutable and attempting to call `put`, trapping the
exception and making the map mutable after this. This approach has not
been taken since the performance benefit has been considered too
negligible. Note that the previous approach also used the exact same
code.

A point of concern may be the additional method handle indirection,
along with `invoke` instead of `invokeExact`. This performance hit is
nevertheless present only once per tag type and the JVM is free to
inline the target method handle if it deems necessary. `invoke` is also
not a problem, since the performance difference between it and
`invokeExact` is negligible. `invokeExact` is still preferable since it
avoids additional indirection due to type conversions, but this little
nanosecond difference shouldn't matter.

All in all, this change is more for personal satisfaction over the
previous solution: the performance benefit of this solution is extremely
small, but being able to go back to a coremod-free CraftTweaker is
definitely something I wanted to try for a long time.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>