Provide Path access to scripts and other directories
Also use the game directory variable to ensure proper resolution,
instead of winging it with relative paths. The end result should match
regardless of the method chosen, but this should be more secure in case
some launchers decide to be different and map the current running
directory in a different place than the game directory.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>