Rewrite script loading API features
Running scripts now involves creating an instance of `IScriptRun` with
the appropriate configuration and source files. The created object can
then be executed right after creation or at a second time, if any
additional operations need to be performed.

The concept of "current run" has been modified to imply a run that is
currently on-going, not the last run that was carried out. `IAction`s
should consider this when updating to the new API. Moreover,
`LoaderActions` have been replaced by `IScriptRunInfo`s, which allows
users to access run information for the current run without being able
to modify it on the fly. Previous runs access is also possible, but
currently restricted to the implementation.

Script run modules can be configured through
`IScriptRunModuleConfigurator`s. Each `IScriptLoader` must have an
associated script run configurator. The default script configurator will
also be provided to ease development and usage of custom loaders.

`IPreprocessor`s and `IAction`s have also been updated to match the new
API. Namely, static access has been removed and preprocessors should now
operate solely based on the data given to them in the `apply` method.
Actions, on the other hand, get to use the updated `IScriptLoadSource`
for checking applicability.

Last but not least, a new `RunKind` enumeration has been added, to allow
users to properly specify if a run should be a formatting, syntax check,
or executing run. It also allows extensibility in case new run types are
added and a more object-oriented approach in the form of `ScriptRunner`
and its subclasses.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>