Rework module registration logic
This allows us to perform operations that need to occur before and after
a module registration in the proper order without having to rely on
plugins to properly implement stuff themselves. While not the best
solution, it is the best course of action given the current state of
ZenCode and the requirements.

Given also the fact that basically nobody else is going to use this API,
it seems like a safe bet that this API is unlikely to require further
changes.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>