Have `<constant>` be available on all loaders
The values that can be grabbed by this bracket expression change
depending on the loader based on the data available in the enum bracket
registry. For this reason, locking out the bracket handler entirely to
the crafttweaker loader only is not only counterintuitive, but it also
does not make a lot of sense.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>